Moves a lot of update conversion logic to ConverterService
