+++ src/test/scala/no/ndla/learningpathapi/TestEnvironment.scala
@@ -37,7 +37,7 @@ trait TestEnvironment
     with ImageApiClientComponent
     with ArticleImportClient
     with MigrationApiClient
-    with ConverterServiceComponent
+    with ConverterService
     with Elastic4sClient
     with DatasourceComponent
     with MockitoSugar
