+++ src/test/scala/no/ndla/learningpathapi/service/ConverterServiceTest.scala
@@ -283,4 +283,56 @@ class ConverterServiceTest extends UnitSuite with UnitTestEnvironment {
       EmbedUrl(externalUrl, "nb", EmbedType.OEmbed))
   }
 
+  test("That mergeLanguageFields returns original list when updated is empty") {
+    val existing = Seq(Title("Tittel 1", "nb"), Title("Tittel 2", "nn"), Title("Tittel 3", "unknown"))
+    service.mergeLanguageFields(existing, Seq()) should equal(existing)
+  }
+
+  test("That mergeLanguageFields updated the english title only when specified") {
+    val tittel1 = Title("Tittel 1", "nb")
+    val tittel2 = Title("Tittel 2", "nn")
+    val tittel3 = Title("Tittel 3", "en")
+    val oppdatertTittel3 = Title("Title 3 in english", "en")
+
+    val existing = Seq(tittel1, tittel2, tittel3)
+    val updated = Seq(oppdatertTittel3)
+
+    service.mergeLanguageFields(existing, updated) should equal(Seq(tittel1, tittel2, oppdatertTittel3))
+  }
+
+  test("That mergeLanguageFields removes a title that is empty") {
+    val tittel1 = Title("Tittel 1", "nb")
+    val tittel2 = Title("Tittel 2", "nn")
+    val tittel3 = Title("Tittel 3", "en")
+    val tittelToRemove = Title("", "nn")
+
+    val existing = Seq(tittel1, tittel2, tittel3)
+    val updated = Seq(tittelToRemove)
+
+    service.mergeLanguageFields(existing, updated) should equal(Seq(tittel1, tittel3))
+  }
+
+  test("That mergeLanguageFields updates the title with no language specified") {
+    val tittel1 = Title("Tittel 1", "nb")
+    val tittel2 = Title("Tittel 2", "unknown")
+    val tittel3 = Title("Tittel 3", "en")
+    val oppdatertTittel2 = Title("Tittel 2 er oppdatert", "unknown")
+
+    val existing = Seq(tittel1, tittel2, tittel3)
+    val updated = Seq(oppdatertTittel2)
+
+    service.mergeLanguageFields(existing, updated) should equal(Seq(tittel1, tittel3, oppdatertTittel2))
+  }
+
+  test("That mergeLanguageFields also updates the correct description") {
+    val desc1 = Description("Beskrivelse 1", "nb")
+    val desc2 = Description("Beskrivelse 2", "unknown")
+    val desc3 = Description("Beskrivelse 3", "en")
+    val oppdatertDesc2 = Description("Beskrivelse 2 er oppdatert", "unknown")
+
+    val existing = Seq(desc1, desc2, desc3)
+    val updated = Seq(oppdatertDesc2)
+    service.mergeLanguageFields(existing, updated) should equal(Seq(desc1, desc3, oppdatertDesc2))
+  }
+
 }
